Rename: 컴포넌트 폴더명 Pascal Case로 변경
