+++ app/(route)/activity/register/_components/BannerImage/index.tsx
@@ -8,8 +8,8 @@ import type { ActivityEdit } from '@/(route)/activity/edit/[id]/page';
 import { usePostImage } from '@/_hooks/use-post-image';
 
 import type { Activity } from '../../page';
-import CommonModal from '../common-modal';
-import FileInput from '../file-input';
+import CommonModal from '../CommonModal';
+import FileInput from '../FileInput';
 
 export interface BannerImageProps<T> {
   setFormData: Dispatch<SetStateAction<T>>;
