+++ app/(route)/activity/register/_components/CommonModal/index.tsx
