+++ app/(route)/activity/register/_components/FileInput/index.tsx
