+++ app/(route)/activity/register/_components/PriceButtons/index.tsx
