+++ app/(route)/activity/register/_components/ScheduleEditor/index.tsx
